implemented analyzer and its ability to use ZMQ to make requests back to KCEF
