+++ analyzer/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "@nlpjs/basic": "^4.27.0",
     "jsdom": "^22.1.0",
-    "nlpjs": "^0.0.2"
+    "nlpjs": "^0.0.2",
+    "zeromq": "6.0.0-beta.17"
   }
 }
